# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator OpenApi",
        "description": "Test your Apis with a full documentation.",
        "contact": {
            "email": "rayanwork2014@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/sources": {
            "get": {
                "tags": [
                    "Sources"
                ],
                "summary": "Get list of sources",
                "description": "Retrieve a paginated list of all sources.",
                "operationId": "getSourcesList",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SourcesListResource"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/PaginationResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sources"
                ],
                "summary": "Create a new source",
                "description": "Create a new source with a specified name.",
                "operationId": "createSource",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the source",
                                        "type": "string",
                                        "example": "Technology"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Source created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Source created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/sources/{id}": {
            "put": {
                "tags": [
                    "Sources"
                ],
                "summary": "Update an existing source",
                "description": "Update the name or status of a source by its ID.",
                "operationId": "updateSource",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Source ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the source",
                                        "type": "string",
                                        "example": "New Source Name",
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "Status of the source",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Source updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Source updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Source not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sources"
                ],
                "summary": "Delete a source",
                "description": "Delete a source by its ID.",
                "operationId": "deleteSource",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Source ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Source deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Source deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Source not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SourcesListResource": {
                "title": "Source List Resource",
                "description": "Basic details of a source.",
                "properties": {
                    "id": {
                        "description": "Unique identifier of the source.",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the source.",
                        "type": "string",
                        "example": "New York Times"
                    }
                },
                "type": "object"
            },
...
            "PaginationResource": {
                "title": "Pagination Resource",
                "description": "Pagination details for paginated API responses.",
                "properties": {
                    "total": {
                        "description": "Total number of items available.",
                        "type": "integer",
                        "example": 100
                    },
                    "count": {
                        "description": "Number of items on the current page.",
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "description": "Number of items per page.",
                        "type": "integer",
                        "example": 10
                    },
                    "page": {
                        "description": "Current page number.",
                        "type": "integer",
                        "example": 1
                    },
                    "max_page": {
                        "description": "Maximum number of pages available.",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}